Note .dockerignore in the exclusion for credentials
The README indicates that the ghe-* files should be
excluded from git via the .gitignore file. However, from CI, it 
is less likely that users will create a new patch back to their version control with the results of their build, but it is more likely that they'll pack it into a container.

This commit clarifies that the container is an equal (or
larger) risk and should be excluded. This also works (for example) with Podman.

See:
1. https://docs.docker.com/engine/reference/builder/


 

Signed-off-by: Andrew Howden <hello@andrewhowden.com>