Add an illustrative example to Direct Workload Identity Federation
Added an illustrative example of adding another IAM policy binding as I found it very confusing to be granting access to secrets. This issue comes up with trying to use https://github.com/google-github-actions/get-gke-credentials as that can fail with a bit of a misleading error of `google-github-actions/get-gke-credentials failed with: required "container.clusters.get" permission(s)` when the actual required role is `roles/container.clusterViewer`

Also added a ´to´ to fix a grammatical error that made the sentence really hard to read. 

Signed-off-by: Oskari Petas <oskari.petas@gmail.com>